requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
bcrypt==4.1.2
Werkzeug==3.0.1
openai>=1.12.0
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
psycopg2-binary==2.9.9